+++ transformer/__init__.py
@@ -1,10 +1,5 @@
 from transformer.model import Transformer
-from transformer.utils import (
-    DataHandler, MetricMeter, LossPenalizer, EarlyStopper, ModelHandler,
-    )
-
-from transformer.variable import create_variable
-from transformer.database import get_metric_db
-from transformer.cli import parse_cli_args
+from transformer.handler import DataHandler, ModelHandler
+from transformer.monitor import MetricMeter, LossPenalizer, EarlyStopper